+++ CHANGELOG.md
@@ -2,12 +2,20 @@
 
 ## Unreleased
 
+### Changed
+
+- Coder output panel enhancements: All log entries now include timestamps, and you
+  can filter messages by log level in the panel.
+
+### Added
+
 - Update `/openDevContainer` to support all dev container features when hostPath
   and configFile are provided.
 - Add `coder.disableUpdateNotifications` setting to disable workspace template
   update notifications.
-- Coder output panel enhancements: All log entries now include timestamps, and you
-  can filter messages by log level in the panel.
+- Add binary signature verification. This can be disabled with
+  `coder.disableSignatureVerification` if you purposefully run a binary that is
+  not signed by Coder (for example a binary you built yourself).
 
 ## [v1.9.2](https://github.com/coder/vscode-coder/releases/tag/v1.9.2) 2025-06-25